requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
azure-core==1.35.0
azure-storage-blob==12.26.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
distro==1.9.0
ecdsa==0.19.1
environs==9.5.0
fastapi==0.116.1
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
jiter==0.10.0
joblib==1.5.2
marshmallow==3.26.1
milvus-lite==2.5.1
minio==7.2.16
numpy==1.26.4
openai==1.106.1
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pluggy==1.6.0
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.22.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymilvus==2.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
structlog==23.2.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
lxml>=6.0.0
pymupdf>=1.26.0
python-docx>=1.2.0
streamlit>=1.40.0
spacy>=3.7.0
